Properly implement NSS key logging
Down at the C level, instances of the Ruby object `SSLSocket` are merely a
wrapper around the OpenSSL top-level `SSL` struct. We can read this data and
use it to generate a string in [NSS Key Log Format][nss-format].

The components of the `SSL` structure used are:

  - `ssl->s3->client_random`: 32 byte client random data.
  - `ssl->session->master_kye`: 48 byte pre master secret key.

The SSL connection must be established before this data is read.

  [nss-format]: https://developer.mozilla.org/en-US/docs/Mozilla/Projects/NSS/Key_Log_Format