Return nil if SSLv2 is in use
These days, use of SSLv2 should be rare and insecure. Handle this case by
returning a `nil` value from the `.to_keylog` function. Pre master secrets for
SSLv2 connections can be logged using an older version of the NSS Key Log
format, but this can be implemented seperately if needed.

Closes GH-5.